refactor(test): consolidate BadgeManager tests into single file
Merge badge-manager.test.ts (18 tests) and badge-manager.integration.test.ts
(7 tests) into one file, eliminating 5 duplicates while preserving all unique
coverage (20 tests).
