+++ src/main/managers/badge-manager.integration.test.ts
@@ -31,22 +31,23 @@ function createMockWindowManager(): MockWindowManager {
   };
 }
 
-describe("BadgeManager Integration", () => {
+describe("BadgeManager", () => {
   let appLayer: MockAppLayer;
   let imageLayer: MockImageLayer;
   let windowManager: MockWindowManager;
 
   beforeEach(() => {
-    appLayer = createAppLayerMock({ platform: "darwin" });
+    appLayer = createAppLayerMock();
     imageLayer = createImageLayerMock();
     windowManager = createMockWindowManager();
   });
 
-  describe("Badge state via updateBadge()", () => {
-    it("shows no badge for none state", () => {
+  describe("updateBadge (darwin)", () => {
+    it("shows filled circle for all-working state", () => {
       const platformInfo = createMockPlatformInfo({ platform: "darwin" });
+      appLayer = createAppLayerMock({ platform: "darwin" });
 
-      const badgeManager = new BadgeManager(
+      const manager = new BadgeManager(
         platformInfo,
         appLayer,
         imageLayer,
@@ -54,43 +55,103 @@ describe("BadgeManager Integration", () => {
         SILENT_LOGGER
       );
 
-      badgeManager.updateBadge("none");
+      manager.updateBadge("all-working");
+
+      expect(appLayer).toHaveDockBadge("●");
+    });
+
+    it("shows half circle for mixed state", () => {
+      const platformInfo = createMockPlatformInfo({ platform: "darwin" });
+      appLayer = createAppLayerMock({ platform: "darwin" });
+
+      const manager = new BadgeManager(
+        platformInfo,
+        appLayer,
+        imageLayer,
+        windowManager as unknown as WindowManager,
+        SILENT_LOGGER
+      );
+
+      manager.updateBadge("mixed");
+
+      expect(appLayer).toHaveDockBadge("◐");
+    });
+
+    it("clears badge for none state", () => {
+      const platformInfo = createMockPlatformInfo({ platform: "darwin" });
+      appLayer = createAppLayerMock({ platform: "darwin" });
+
+      const manager = new BadgeManager(
+        platformInfo,
+        appLayer,
+        imageLayer,
+        windowManager as unknown as WindowManager,
+        SILENT_LOGGER
+      );
+
+      manager.updateBadge("none");
+
       expect(appLayer).toHaveDockBadge("");
-    });
-
-    it("shows red badge for all-working state", () => {
-      const platformInfo = createMockPlatformInfo({ platform: "darwin" });
-
-      const badgeManager = new BadgeManager(
-        platformInfo,
-        appLayer,
-        imageLayer,
-        windowManager as unknown as WindowManager,
-        SILENT_LOGGER
-      );
-
-      badgeManager.updateBadge("all-working");
-      expect(appLayer).toHaveDockBadge("\u25CF");
-    });
-
-    it("shows mixed badge for mixed state", () => {
-      const platformInfo = createMockPlatformInfo({ platform: "darwin" });
-
-      const badgeManager = new BadgeManager(
-        platformInfo,
-        appLayer,
-        imageLayer,
-        windowManager as unknown as WindowManager,
-        SILENT_LOGGER
-      );
-
-      badgeManager.updateBadge("mixed");
-      expect(appLayer).toHaveDockBadge("\u25D0");
     });
   });
 
-  describe("Windows overlay icon", () => {
-    it("creates image in same ImageLayer that WindowLayer would use for lookup", () => {
+  describe("updateBadge (win32)", () => {
+    it("generates image for all-working state", () => {
+      const platformInfo = createMockPlatformInfo({ platform: "win32" });
+
+      const manager = new BadgeManager(
+        platformInfo,
+        appLayer,
+        imageLayer,
+        windowManager as unknown as WindowManager,
+        SILENT_LOGGER
+      );
+
+      manager.updateBadge("all-working");
+
+      expect(imageLayer).toHaveImages([{ id: "image-1" }]);
+      expect(windowManager.setOverlayIconCalls).toHaveLength(1);
+      expect(windowManager.setOverlayIconCalls[0]?.description).toBe("All workspaces working");
+    });
+
+    it("generates image for mixed state", () => {
+      const platformInfo = createMockPlatformInfo({ platform: "win32" });
+
+      const manager = new BadgeManager(
+        platformInfo,
+        appLayer,
+        imageLayer,
+        windowManager as unknown as WindowManager,
+        SILENT_LOGGER
+      );
+
+      manager.updateBadge("mixed");
+
+      expect(imageLayer).toHaveImages([{ id: "image-1" }]);
+      expect(windowManager.setOverlayIconCalls).toHaveLength(1);
+      expect(windowManager.setOverlayIconCalls[0]?.description).toBe("Some workspaces ready");
+    });
+
+    it("clears overlay for none state", () => {
+      const platformInfo = createMockPlatformInfo({ platform: "win32" });
+
+      const manager = new BadgeManager(
+        platformInfo,
+        appLayer,
+        imageLayer,
+        windowManager as unknown as WindowManager,
+        SILENT_LOGGER
+      );
+
+      manager.updateBadge("none");
+
+      expect(imageLayer).toHaveImages([]);
+      expect(windowManager.setOverlayIconCalls).toHaveLength(1);
+      expect(windowManager.setOverlayIconCalls[0]?.image).toBeNull();
+      expect(windowManager.setOverlayIconCalls[0]?.description).toBe("");
+    });
+
+    it("creates image in shared ImageLayer that WindowLayer would use for lookup", () => {
       const platformInfo = createMockPlatformInfo({ platform: "win32" });
       appLayer = createAppLayerMock({ platform: "win32" });
       const sharedImageLayer = createImageLayerMock();
@@ -102,7 +163,7 @@ describe("BadgeManager Integration", () => {
         },
       };
 
-      const badgeManager = new BadgeManager(
+      const manager = new BadgeManager(
         platformInfo,
         appLayer,
         sharedImageLayer,
@@ -110,7 +171,7 @@ describe("BadgeManager Integration", () => {
         SILENT_LOGGER
       );
 
-      badgeManager.updateBadge("all-working");
+      manager.updateBadge("all-working");
 
       expect(capturedImageHandle).not.toBeNull();
       expect(sharedImageLayer).toHaveImage(capturedImageHandle!.id, {
@@ -118,12 +179,14 @@ describe("BadgeManager Integration", () => {
         size: { width: 16, height: 16 },
       });
     });
+  });
 
-    it("generates image for mixed state", () => {
-      const platformInfo = createMockPlatformInfo({ platform: "win32" });
-      appLayer = createAppLayerMock({ platform: "win32" });
+  describe("updateBadge (linux)", () => {
+    it("sets badge count to 1 for all-working state", () => {
+      const platformInfo = createMockPlatformInfo({ platform: "linux" });
+      appLayer = createAppLayerMock({ platform: "linux" });
 
-      const badgeManager = new BadgeManager(
+      const manager = new BadgeManager(
         platformInfo,
         appLayer,
         imageLayer,
@@ -131,18 +194,16 @@ describe("BadgeManager Integration", () => {
         SILENT_LOGGER
       );
 
-      badgeManager.updateBadge("mixed");
+      manager.updateBadge("all-working");
 
-      expect(imageLayer).toHaveImages([{ id: "image-1" }]);
-      expect(windowManager.setOverlayIconCalls).toHaveLength(1);
-      expect(windowManager.setOverlayIconCalls[0]?.description).toBe("Some workspaces ready");
+      expect(appLayer).toHaveBadgeCount(1);
     });
 
-    it("generates image for all-working state", () => {
-      const platformInfo = createMockPlatformInfo({ platform: "win32" });
-      appLayer = createAppLayerMock({ platform: "win32" });
+    it("sets badge count to 1 for mixed state", () => {
+      const platformInfo = createMockPlatformInfo({ platform: "linux" });
+      appLayer = createAppLayerMock({ platform: "linux" });
 
-      const badgeManager = new BadgeManager(
+      const manager = new BadgeManager(
         platformInfo,
         appLayer,
         imageLayer,
@@ -150,19 +211,139 @@ describe("BadgeManager Integration", () => {
         SILENT_LOGGER
       );
 
-      badgeManager.updateBadge("all-working");
+      manager.updateBadge("mixed");
+
+      expect(appLayer).toHaveBadgeCount(1);
+    });
+
+    it("clears badge for none state", () => {
+      const platformInfo = createMockPlatformInfo({ platform: "linux" });
+      appLayer = createAppLayerMock({ platform: "linux" });
+
+      const manager = new BadgeManager(
+        platformInfo,
+        appLayer,
+        imageLayer,
+        windowManager as unknown as WindowManager,
+        SILENT_LOGGER
+      );
+
+      manager.updateBadge("none");
+
+      expect(appLayer).toHaveBadgeCount(0);
+    });
+  });
+
+  describe("generateBadgeImage", () => {
+    it("creates a 16x16 bitmap image for all-working", () => {
+      const platformInfo = createMockPlatformInfo({ platform: "win32" });
+
+      const manager = new BadgeManager(
+        platformInfo,
+        appLayer,
+        imageLayer,
+        windowManager as unknown as WindowManager,
+        SILENT_LOGGER
+      );
+
+      manager.updateBadge("all-working");
+
+      expect(imageLayer).toHaveImage("image-1", { size: { width: 16, height: 16 } });
+    });
+
+    it("creates different images for different states", () => {
+      const platformInfo = createMockPlatformInfo({ platform: "win32" });
+
+      const manager = new BadgeManager(
+        platformInfo,
+        appLayer,
+        imageLayer,
+        windowManager as unknown as WindowManager,
+        SILENT_LOGGER
+      );
+
+      manager.updateBadge("all-working");
+      manager.updateBadge("mixed");
+
+      expect(imageLayer).toHaveImages([{ id: "image-1" }, { id: "image-2" }]);
+    });
+
+    it("creates non-empty image for mixed state", () => {
+      const platformInfo = createMockPlatformInfo({ platform: "win32" });
+
+      const manager = new BadgeManager(
+        platformInfo,
+        appLayer,
+        imageLayer,
+        windowManager as unknown as WindowManager,
+        SILENT_LOGGER
+      );
+
+      manager.updateBadge("mixed");
+
+      expect(imageLayer).toHaveImage("image-1", { isEmpty: false });
+    });
+
+    it("creates non-empty image for all-working state", () => {
+      const platformInfo = createMockPlatformInfo({ platform: "win32" });
+
+      const manager = new BadgeManager(
+        platformInfo,
+        appLayer,
+        imageLayer,
+        windowManager as unknown as WindowManager,
+        SILENT_LOGGER
+      );
+
+      manager.updateBadge("all-working");
+
+      expect(imageLayer).toHaveImage("image-1", { isEmpty: false });
+    });
+  });
+
+  describe("image caching", () => {
+    it("reuses cached images for same state", () => {
+      const platformInfo = createMockPlatformInfo({ platform: "win32" });
+
+      const manager = new BadgeManager(
+        platformInfo,
+        appLayer,
+        imageLayer,
+        windowManager as unknown as WindowManager,
+        SILENT_LOGGER
+      );
+
+      manager.updateBadge("all-working");
+      manager.updateBadge("all-working");
+      manager.updateBadge("all-working");
 
       expect(imageLayer).toHaveImages([{ id: "image-1" }]);
-      expect(windowManager.setOverlayIconCalls).toHaveLength(1);
-      expect(windowManager.setOverlayIconCalls[0]?.description).toBe("All workspaces working");
+      expect(windowManager.setOverlayIconCalls).toHaveLength(3);
+    });
+
+    it("creates new images for different states", () => {
+      const platformInfo = createMockPlatformInfo({ platform: "win32" });
+
+      const manager = new BadgeManager(
+        platformInfo,
+        appLayer,
+        imageLayer,
+        windowManager as unknown as WindowManager,
+        SILENT_LOGGER
+      );
+
+      manager.updateBadge("all-working");
+      manager.updateBadge("mixed");
+
+      expect(imageLayer).toHaveImages([{ id: "image-1" }, { id: "image-2" }]);
     });
   });
 
   describe("dispose", () => {
-    it("clears badge and releases images on dispose", () => {
-      const platformInfo = createMockPlatformInfo({ platform: "darwin" });
+    it("releases all cached images", () => {
+      const platformInfo = createMockPlatformInfo({ platform: "win32" });
 
-      const badgeManager = new BadgeManager(
+      const manager = new BadgeManager(
         platformInfo,
         appLayer,
         imageLayer,
@@ -170,13 +351,75 @@ describe("BadgeManager Integration", () => {
         SILENT_LOGGER
       );
 
-      badgeManager.updateBadge("all-working");
-      expect(appLayer).toHaveDockBadge("\u25CF");
+      manager.updateBadge("all-working");
+      manager.updateBadge("mixed");
 
-      badgeManager.dispose();
+      expect(imageLayer).toHaveImages([{ id: "image-1" }, { id: "image-2" }]);
 
-      // Badge should be cleared
+      manager.dispose();
+
+      expect(imageLayer).toHaveImages([]);
+    });
+
+    it("clears overlay on dispose (win32)", () => {
+      const platformInfo = createMockPlatformInfo({ platform: "win32" });
+
+      const manager = new BadgeManager(
+        platformInfo,
+        appLayer,
+        imageLayer,
+        windowManager as unknown as WindowManager,
+        SILENT_LOGGER
+      );
+
+      manager.updateBadge("all-working");
+      const callsAfterUpdate = windowManager.setOverlayIconCalls.length;
+      expect(callsAfterUpdate).toBeGreaterThan(0);
+
+      manager.dispose();
+
+      const lastCall = windowManager.setOverlayIconCalls.at(-1);
+      expect(lastCall?.image).toBeNull();
+    });
+
+    it("clears badge on dispose (darwin)", () => {
+      const platformInfo = createMockPlatformInfo({ platform: "darwin" });
+      appLayer = createAppLayerMock({ platform: "darwin" });
+
+      const manager = new BadgeManager(
+        platformInfo,
+        appLayer,
+        imageLayer,
+        windowManager as unknown as WindowManager,
+        SILENT_LOGGER
+      );
+
+      manager.updateBadge("all-working");
+      expect(appLayer).toHaveDockBadge("●");
+
+      manager.dispose();
+
       expect(appLayer).toHaveDockBadge("");
+    });
+
+    it("is idempotent - can be called multiple times safely", () => {
+      const platformInfo = createMockPlatformInfo({ platform: "win32" });
+
+      const manager = new BadgeManager(
+        platformInfo,
+        appLayer,
+        imageLayer,
+        windowManager as unknown as WindowManager,
+        SILENT_LOGGER
+      );
+
+      manager.updateBadge("all-working");
+
+      manager.dispose();
+      manager.dispose();
+      manager.dispose();
+
+      expect(imageLayer).toHaveImages([]);
     });
   });
 });
